[feat] attraction 지역 정보 불러오기 service 추가

+++ bonvoyageW/src/main/java/com/bonvoyage/domain/attraction/repository/AttractionDetailRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AttractionDetailRepository extends JpaRepository<AttractionDetailEntity, Integer> {
+public interface AttractionDetailRepository extends JpaRepository<AttractionDetailEntity, Long> {
 }
